fix(desktop): keep export/session rows live while the palette is open
Freeze-at-open captured messages/sessions at build time, so export/save
while the palette stayed open used a stale snapshot, and session catalog
rows stopped updating on background create/rename. Base commands still
freeze per open/close with run() reading the latest options ref; session
rows rebuild on visibleSessions + activeId only. Extract pure builders for
unit tests and co-migrate contracts. Refs #1045.
