+++ apps/desktop/src/main/__tests__/command-palette-plan-reminder-contract.test.ts
@@ -8,7 +8,7 @@ const REPO_ROOT = resolve(import.meta.dirname, '../../../../..');
 
 describe('command palette plan reminder contract', () => {
   it('exposes a direct action for starting a new plan reminder', async () => {
-    const src = await readFile(resolve(REPO_ROOT, 'apps/desktop/src/renderer/command-palette.tsx'), 'utf8');
+    const src = await readFile(resolve(REPO_ROOT, 'apps/desktop/src/renderer/command-palette-commands.ts'), 'utf8');
 
     assert.match(src, /onStartPlanReminder\?\(\): void/);
     assert.match(src, /id:\s*'action:new-plan-reminder'/);
@@ -23,7 +23,8 @@ describe('command palette plan reminder contract', () => {
 
     assert.match(main, /function\s+openPlanReminderForm\(\)/);
     assert.match(main, /setNavSelection\(\{\s*section:\s*'automations'\s*\}\)/);
-    assert.match(main, /onStartPlanReminder:\s*openPlanReminderForm/);
+    // #1045: run() reads openPlanReminderForm from the live options ref.
+    assert.match(main, /onStartPlanReminder:\s*\(\)\s*=>\s*optionsRef\.current\.openPlanReminderForm\(\)/);
     assert.match(main, /querySelector<HTMLInputElement>\('\[data-maka-plan-title-input="true"\]'\)/);
     assert.match(ui, /data-maka-plan-title-input="true"/);
   });
