+++ apps/desktop/src/main/__tests__/command-palette-session-navigation-contract.test.ts
@@ -5,17 +5,25 @@ import { readRendererShellCombinedSource } from './renderer-shell-source-helpers
 describe('command palette session navigation contract', () => {
   it('routes normal session commands back to the chat surface before selecting the session', async () => {
     const main = await readRendererShellCombinedSource();
-    const paletteBlock = main.match(/return buildCommandList\(\{[\s\S]*?onNewChat:/)?.[0] ?? '';
+    // #1045: session rows live in buildAppShellSessionCommands (not the frozen
+    // base buildCommandList). Same invariant: openSessionInChat, not setActiveId.
+    const sessionBlock =
+      main.match(/export function buildAppShellSessionCommands\([\s\S]*?\n\}/)?.[0] ?? '';
 
     assert.match(
-      paletteBlock,
+      sessionBlock,
       /onSelectSession: \(sessionId\) => \{[\s\S]*openSessionInChat\(sessionId\);[\s\S]*\}/,
       'ordinary Command Palette session hits must switch modules back to Chat before selecting the session',
     );
     assert.doesNotMatch(
-      paletteBlock,
+      sessionBlock,
       /onSelectSession: setActiveId/,
       'passing setActiveId directly makes palette session hits invisible from Plan / Daily Review / Skills modules',
+    );
+    assert.match(
+      sessionBlock,
+      /optionsRef\.current\.openSessionInChat\(sessionId\)/,
+      'session select must read openSessionInChat from the live options ref at run time',
     );
   });
 });
