+++ apps/desktop/src/renderer/app-shell-command-actions.ts
@@ -1,3 +1,4 @@
+import { useMemo, useRef } from 'react';
 import type {
   DailyReviewSummary,
   LlmConnection,
@@ -10,10 +11,11 @@ import type {
 } from '@maka/core';
 import { formatDailyReviewMarkdown } from '@maka/ui';
 import type { NavSelection } from '@maka/ui';
-import { buildCommandList } from './command-palette';
-import { renderConversationMarkdown } from './conversation-markdown';
-import { dailyReviewActionErrorMessage } from './daily-review-actions';
-import { commandPaletteActionErrorMessage, commandPaletteConnectionTestFailureMessage } from './app-shell-copy';
+import { buildCommandList, buildSessionCommands } from './command-palette-commands.js';
+import type { Command } from './command-palette-types.js';
+import { renderConversationMarkdown } from './conversation-markdown.js';
+import { dailyReviewActionErrorMessage } from './daily-review-actions.js';
+import { commandPaletteActionErrorMessage, commandPaletteConnectionTestFailureMessage } from './app-shell-copy.js';
 
 type ToastApi = {
   success(title: string, description?: string): void;
@@ -72,63 +74,39 @@ export interface AppShellCommandListOptions {
   toastApi: ToastApi;
 }
 
-export function buildAppShellCommandList(options: AppShellCommandListOptions): ReturnType<typeof buildCommandList> {
-  const {
-    activeId,
-    activePermissionMode,
-    captureComposerImportOwner,
-    closePalette,
-    composerRef,
-    connections,
-    createSession,
-    dailyReviewBridge,
-    defaultConnection,
-    handleQuickChatSubmit,
-    isComposerImportOwnerActive,
-    messages,
-    openHelp,
-    openPlanReminderForm,
-    openProjectFolder,
-    openSessionInChat,
-    openSettings,
-    openSettingsSection,
-    openSkillsFolder,
-    openWorkspaceFolder,
-    refreshConnections,
-    saveDailyReviewMarkdown,
-    sessions,
-    setNavSelection,
-    setPermissionMode,
-    setThemePref,
-    themePref,
-    toastApi,
-    visibleSessions,
-  } = options;
+export function buildAppShellCommandList(
+  optionsRef: RefBox<AppShellCommandListOptions>,
+): ReturnType<typeof buildCommandList> {
+  // #1045: useAppShellCommands freezes this list per palette open/close
+  // transition. List-SHAPING fields (which rows exist, labels, hints) come
+  // from the build-time snapshot below; every value a command touches at RUN
+  // time is dereferenced from the ref inside the callback, so the frozen list
+  // still acts on current data (same stable-ref pattern as
+  // openSessionInChatRef in app-shell.tsx).
+  const options = optionsRef.current;
 
   return buildCommandList({
-    sessions: visibleSessions,
-    activeSessionId: activeId,
-    themePref: themePref,
-    connections: connections,
-    defaultSlug: defaultConnection,
-    onSelectSession: (sessionId) => {
-      openSessionInChat(sessionId);
-    },
-    onNewChat: () => createSession(),
+    activeSessionId: options.activeId,
+    themePref: options.themePref,
+    connections: options.connections,
+    defaultSlug: options.defaultConnection,
+    onNewChat: () => optionsRef.current.createSession(),
     onStartDeepResearch: async () => {
+      const { handleQuickChatSubmit } = optionsRef.current;
       await handleQuickChatSubmit('', 'deep_research');
     },
-    onStartPlanReminder: openPlanReminderForm,
-    onOpenSettings: openSettings,
-    onOpenSettingsSection: (section) => openSettingsSection(section),
+    onStartPlanReminder: () => optionsRef.current.openPlanReminderForm(),
+    onOpenSettings: () => optionsRef.current.openSettings(),
+    onOpenSettingsSection: (section) => optionsRef.current.openSettingsSection(section),
     // PR-UX-POLISH-1 commit 4 (WAWQAQ `e0dbad11` + kenji `2844f64f`):
     // use the openHelp callback returned by useKeyboardHelp directly,
     // instead of dispatching a synthetic KeyboardEvent. Same effect,
     // clearer intent, and avoids the foot-gun where a typed `?` in a
     // text input would be swallowed by the global keydown listener.
-    onOpenShortcuts: openHelp,
-    onSetTheme: setThemePref,
+    onOpenShortcuts: () => optionsRef.current.openHelp(),
+    onSetTheme: (next) => optionsRef.current.setThemePref(next),
     onTestConnection: async (slug) => {
+      const { connections, refreshConnections, toastApi } = optionsRef.current;
       try {
         const result = await window.maka.connections.test(slug);
         const conn = connections.find((c) => c.slug === slug);
@@ -150,6 +128,7 @@ export function buildAppShellCommandList(options: AppShellCommandListOptions): R
       }
     },
     onSetDefaultConnection: async (slug) => {
+      const { connections, refreshConnections, toastApi } = optionsRef.current;
       try {
         await window.maka.connections.setDefault(slug);
         await refreshConnections();
@@ -163,15 +142,17 @@ export function buildAppShellCommandList(options: AppShellCommandListOptions): R
       }
     },
     onOpenWorkspace: async () => {
-      await openWorkspaceFolder();
+      await optionsRef.current.openWorkspaceFolder();
     },
-    onOpenProjectFolder: () => openProjectFolder(),
-    onOpenSkillsFolder: () => openSkillsFolder(),
+    onOpenProjectFolder: () => optionsRef.current.openProjectFolder(),
+    onOpenSkillsFolder: () => optionsRef.current.openSkillsFolder(),
     onSelectModule: (selection) => {
+      const { closePalette, setNavSelection } = optionsRef.current;
       setNavSelection(selection);
       closePalette();
     },
     onExportActiveConversation: async () => {
+      const { activeId, messages, sessions, toastApi } = optionsRef.current;
       if (!activeId) return;
       const session = sessions.find((s) => s.id === activeId);
       const markdown = renderConversationMarkdown(session?.name ?? '新建对话', messages);
@@ -186,6 +167,7 @@ export function buildAppShellCommandList(options: AppShellCommandListOptions): R
       }
     },
     onSaveActiveConversationToFile: async () => {
+      const { activeId, messages, sessions, toastApi } = optionsRef.current;
       if (!activeId) return;
       const session = sessions.find((s) => s.id === activeId);
       const sessionName = session?.name ?? '新建对话';
@@ -223,6 +205,7 @@ export function buildAppShellCommandList(options: AppShellCommandListOptions): R
       }
     },
     onOpenLocalMemoryFile: async () => {
+      const { toastApi } = optionsRef.current;
       try {
         const result = await window.maka.memory.openFile();
         if (!result.ok) {
@@ -236,6 +219,7 @@ export function buildAppShellCommandList(options: AppShellCommandListOptions): R
       }
     },
     onOpenWorkspaceInstructionsFile: async () => {
+      const { toastApi } = optionsRef.current;
       try {
         // PR-CMD-PALETTE-OPEN-WORKSPACE-INSTRUCTIONS-0: open the
         // first available workspace instruction file. If none are
@@ -261,9 +245,10 @@ export function buildAppShellCommandList(options: AppShellCommandListOptions): R
         );
       }
     },
-    onSetPermissionMode: (mode) => setPermissionMode(mode),
-    activePermissionMode: activePermissionMode,
+    onSetPermissionMode: (mode) => optionsRef.current.setPermissionMode(mode),
+    activePermissionMode: options.activePermissionMode,
     onCopyTodayDailyReview: async () => {
+      const { dailyReviewBridge, toastApi } = optionsRef.current;
       try {
         const summary = await dailyReviewBridge.fetchDay(0, 1);
         const markdown = formatDailyReviewMarkdown(summary, '今天');
@@ -280,6 +265,13 @@ export function buildAppShellCommandList(options: AppShellCommandListOptions): R
       }
     },
     onPasteTodayDailyReviewIntoComposer: async () => {
+      const {
+        captureComposerImportOwner,
+        composerRef,
+        dailyReviewBridge,
+        isComposerImportOwnerActive,
+        toastApi,
+      } = optionsRef.current;
       const owner = captureComposerImportOwner();
       if (!owner.sessionId) return;
       try {
@@ -301,6 +293,7 @@ export function buildAppShellCommandList(options: AppShellCommandListOptions): R
       }
     },
     onSaveTodayDailyReviewToFile: async () => {
+      const { dailyReviewBridge, saveDailyReviewMarkdown, toastApi } = optionsRef.current;
       try {
         const summary = await dailyReviewBridge.fetchDay(0, 1);
         const markdown = formatDailyReviewMarkdown(summary, '今天');
@@ -313,6 +306,7 @@ export function buildAppShellCommandList(options: AppShellCommandListOptions): R
       }
     },
     onCopyEnvSummary: async () => {
+      const { toastApi } = optionsRef.current;
       try {
         const info = await window.maka.app.info();
         const platformPretty =
@@ -350,6 +344,7 @@ export function buildAppShellCommandList(options: AppShellCommandListOptions): R
       }
     },
     onTestNetworkProxy: async () => {
+      const { toastApi } = optionsRef.current;
       try {
         // PR-CMD-PALETTE-NETWORK-PROXY-TEST-0: surface the
         // proxy test result via toast so a user debugging a
@@ -371,4 +366,49 @@ export function buildAppShellCommandList(options: AppShellCommandListOptions): R
       }
     },
   });
+}
+
+export function buildAppShellSessionCommands(
+  optionsRef: RefBox<AppShellCommandListOptions>,
+): ReturnType<typeof buildSessionCommands> {
+  const options = optionsRef.current;
+  return buildSessionCommands({
+    sessions: options.visibleSessions,
+    activeSessionId: options.activeId,
+    onSelectSession: (sessionId) => {
+      optionsRef.current.openSessionInChat(sessionId);
+    },
+  });
+}
+
+/**
+ * #1045: the palette's command list keeps a stable identity while it is open.
+ * app-shell rebuilds commandOptions on every render (streaming ticks
+ * included), so the base commands are built once per open/close transition —
+ * their run() closures dereference the latest options through the ref, so the
+ * frozen list still acts on current data. Session rows are derived separately,
+ * memoized on the visible session catalog + active session only: background
+ * session creates/renames stay live while the palette is open, without
+ * reintroducing per-tick rebuilds (visibleSessions is itself memoized in
+ * app-shell, so rows rebuild only on real catalog changes).
+ */
+export function useAppShellCommands(
+  paletteOpen: boolean,
+  commandOptions: AppShellCommandListOptions,
+): Command[] {
+  const optionsRef = useRef(commandOptions);
+  optionsRef.current = commandOptions;
+  const baseCommands = useMemo(
+    () => buildAppShellCommandList(optionsRef),
+    [paletteOpen],
+  );
+  const { activeId, visibleSessions } = commandOptions;
+  const sessionCommands = useMemo(
+    () => buildAppShellSessionCommands(optionsRef),
+    [paletteOpen, visibleSessions, activeId],
+  );
+  return useMemo(
+    () => [...baseCommands, ...sessionCommands],
+    [baseCommands, sessionCommands],
+  );
 }
